+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.1
+### Added
+- algod.send_transaction(): sends SignedTransaction
+
+### Changed
+- algod.send_raw_transaction(): sends base64 encoded transaction
+- Multisig.get_account_from_sig() is now Multisig.get_multisig_account
+
 ## 1.0.0
 ### Added
 - SDK released